+++ payment_method_gateway.go
@@ -2,10 +2,26 @@ package braintree
 
 import (
 	"context"
+	"encoding/xml"
 )
 
 type PaymentMethodGateway struct {
 	*Braintree
+}
+
+type PaymentMethodRequest struct {
+	XMLName            xml.Name                     `xml:"payment-method"`
+	CustomerId         string                       `xml:"customer-id,omitempty"`
+	Token              string                       `xml:"token,omitempty"`
+	PaymentMethodNonce string                       `xml:"payment-method-nonce,omitempty"`
+	Options            *PaymentMethodRequestOptions `xml:"options,omitempty"`
+}
+
+type PaymentMethodRequestOptions struct {
+	MakeDefault                   bool   `xml:"make-default,omitempty"`
+	FailOnDuplicatePaymentMethod  bool   `xml:"fail-on-duplicate-payment-method,omitempty"`
+	VerifyCard                    *bool  `xml:"verify-card,omitempty"`
+	VerificationMerchantAccountId string `xml:"verification-merchant-account-id,omitempty"`
 }
 
 func (g *PaymentMethodGateway) Create(ctx context.Context, paymentMethodRequest *PaymentMethodRequest) (PaymentMethod, error) {
